requirements.txt
aiohttp==3.11.12
backoff==2.2.1
besser==2.5.1
besser-agentic-framework==3.0.1
dateparser==1.1.8
Jinja2==3.1.4
langdetect==1.0.9
numpy==1.24.3
nltk==3.9.1
numpy==1.26.1
openai==1.30.5
pandas==2.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pyvis==0.3.2
requests==2.31.0
snowballstemmer==2.2.0
sqlalchemy==2.0.29
streamlit==1.40.0
streamlit-antd-components==0.3.2
text2num==2.5.0
textX==4.0.1
tree-of-thoughts-llm==0.1.0
websocket-client==1.6.4
websockets==11.0.3
